Refactor src/app.ts into separate files src/app/*
Caleydo/cohort#693
