+++ src/Provenance/General.ts
@@ -1,5 +1,5 @@
 import { ActionMetaData, ActionUtils, ICmdResult, IObjectRef, ObjectRefUtils } from 'tdp_core';
-import { CohortApp } from '../app';
+import { CohortApp } from '../app/CohortApp';
 import { IElementProvJSON, IElementProvJSONCohort } from '../CohortInterfaces';
 import { log } from '../util';
 import { IEntitySourceConfig } from '../utilIdTypes';
