+++ src/index.ts
@@ -1,8 +1,9 @@
+export * from './app';
 export * from './data';
 export * from './Overview';
 export * from './Taskview';
 
-export * from './app';
+export * from './app/Coral';
 export * from './Cohort';
 export * from './CohortInterfaces';
 export * from './CohortRepresentations';
